[wasm][debugger][tests] Move stepping tests to a separate file
