+++ src/mono/wasm/debugger/DebuggerTestSuite/Support.cs
@@ -480,14 +480,15 @@ namespace DebuggerTests
         internal async Task<JObject> StepAndCheck(StepKind kind, string script_loc, int line, int column, string function_name,
             Func<JObject, Task> wait_for_event_fn = null, Action<JToken> locals_fn = null, int times = 1)
         {
+            string method = (kind == StepKind.Resume ? "Debugger.resume" : $"Debugger.step{kind}");
             for (int i = 0; i < times - 1; i++)
             {
-                await SendCommandAndCheck(null, $"Debugger.step{kind.ToString()}", null, -1, -1, null);
+                await SendCommandAndCheck(null, method, null, -1, -1, null);
             }
 
             // Check for method/line etc only at the last step
             return await SendCommandAndCheck(
-                null, $"Debugger.step{kind.ToString()}", script_loc, line, column, function_name,
+                null, method, script_loc, line, column, function_name,
                 wait_for_event_fn: wait_for_event_fn,
                 locals_fn: locals_fn);
         }
@@ -510,12 +511,15 @@ namespace DebuggerTests
             }
 
             var wait_res = await ctx.insp.WaitFor(waitForEvent);
-
+            JToken top_frame = wait_res["callFrames"]?[0];
             if (function_name != null)
-                Assert.Equal(function_name, wait_res["callFrames"]?[0]?["functionName"]?.Value<string>());
+            {
+                AssertEqual(function_name, wait_res["callFrames"]?[0]?["functionName"]?.Value<string>(), top_frame?.ToString());
+            }
 
-            if (script_loc != null)
-                CheckLocation(script_loc, line, column, ctx.scripts, wait_res["callFrames"][0]["location"]);
+            Console.WriteLine (top_frame);
+            if (script_loc != null && line >= 0)
+                CheckLocation(script_loc, line, column, ctx.scripts, top_frame["location"]);
 
             if (wait_for_event_fn != null)
                 await wait_for_event_fn(wait_res);
@@ -1094,6 +1098,7 @@ namespace DebuggerTests
     {
         Into,
         Over,
-        Out
+        Out,
+        Resume
     }
 }
